rank chart

+++ functions.py
@@ -205,12 +205,14 @@ def categories(df):
 
 def uno_reverse(gc, df, sheet_key, nominee):
     try:
-        filtered_df = df[(df["drinker_name"] == nominee) & (df["nomination_completed_date"] == "Not Completed")]
+        filtered_df = df[(df["drinker_name"] == nominee) & (
+            df["nomination_completed_date"] == "Not Completed")]
         last_record_index = filtered_df.index[-1]
 
-        df.at[last_record_index,"drinker_name"] = filtered_df.nominator_name[last_record_index]
-        #commented out to prevent uno reverse of an uno reverse
-        # df.at[last_record_index, "nominator_name"] = nominee 
+        df.at[last_record_index,
+              "drinker_name"] = filtered_df.nominator_name[last_record_index]
+        # commented out to prevent uno reverse of an uno reverse
+        # df.at[last_record_index, "nominator_name"] = nominee
         df.at[last_record_index, "drink_type"] = "uno reverse"
 
     except IndexError as e:
@@ -238,7 +240,10 @@ def uno_reverse(gc, df, sheet_key, nominee):
         print("An error occurred:", e)
         return None
 
+
 def build_rank_df(gameweek_df, current_week):
-    last_10 = gameweek_df[gameweek_df.event >= current_week -10].iloc[:,[0,2,4]]
-    last_10['rank'] = last_10.groupby(['event'])['total_points'].rank(ascending=False)
+    last_10 = gameweek_df[gameweek_df.event >=current_week - 10].iloc[:, [0, 2, 4]]
+    # last_10 = gameweek_df[gameweek_df.event<=10].iloc[:,[0,2,4]]
+    last_10['rank'] = last_10.groupby(
+        ['event'])['total_points'].rank(ascending=False).astype(int)
     return last_10
